# specs/001-connect-review-backend/contracts/review-api.openapi.yaml
openapi: 3.1.0
info:
  title: Regulations Document Platform Review API
  version: 1.0.0-mvp
  description: >-
    智能审查领域契约。/api/docs 路径是现有文档 API 的兼容扩展；
    /api/review 路径承载审查领域资源。通用 /api/chat 契约保持不变，未在此重复声明。
servers:
  - url: http://127.0.0.1:8002
...
paths:
...
  /api/review/batches:
    post:
      tags: [Review Batches]
      operationId: createReviewBatch
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateBatchRequest'}
      responses:
        '201':
          description: Batch created
          content:
            application/json: {schema: {$ref: '#/components/schemas/ReviewBatch'}}
        '422': {$ref: '#/components/responses/ValidationError'}
    get:
      tags: [Review Batches]
      operationId: listReviewBatches
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Batch page
          content:
            application/json: {schema: {$ref: '#/components/schemas/BatchPage'}}
  /api/review/batches/{batch_id}:
    get:
      tags: [Review Batches]
      operationId: getReviewBatch
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Batch and current document membership
          content:
            application/json: {schema: {$ref: '#/components/schemas/ReviewBatch'}}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Review Batches]
      operationId: updateReviewBatch
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/IfMatchRevision'
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/UpdateBatchRequest'}}
      responses:
        '200':
          description: Updated batch
          content:
            application/json: {schema: {$ref: '#/components/schemas/ReviewBatch'}}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/review/batches/{batch_id}/documents:
    post:
      tags: [Review Batches]
      operationId: addBatchDocument
      parameters:
        - $ref: '#/components/parameters/BatchId'
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/AddBatchDocumentRequest'}}
      responses:
        '201':
          description: Membership created; version is explicit and immutable
          content:
            application/json: {schema: {$ref: '#/components/schemas/BatchDocument'}}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/review/batches/{batch_id}/documents/{membership_id}:
    delete:
      tags: [Review Batches]
      operationId: removeBatchDocument
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - name: membership_id
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Uuid'}
      responses:
        '204': {description: Removed before analysis snapshot exists}
        '409': {$ref: '#/components/responses/Conflict'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/review/batches/{batch_id}/template-suggestions:
    get:
      tags: [Review Batches]
      operationId: listTemplateSuggestions
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Suggestions only; client must explicitly select a version
          content:
            application/json: {schema: {$ref: '#/components/schemas/TemplateSuggestionPage'}}
...
  /api/review/templates:
    post:
      tags: [Review Rules]
      operationId: createTemplateVersion
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/CreateTemplateVersionRequest'}}
      responses:
        '201':
          description: Draft template version referencing an existing template source version
          content:
            application/json: {schema: {$ref: '#/components/schemas/TemplateVersion'}}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
    get:
      tags: [Review Rules]
      operationId: listTemplates
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: category
          in: query
          schema: {type: string}
        - name: q
          in: query
          schema: {type: string, maxLength: 200}
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/PublishStatus'}
      responses:
        '200':
          description: Template family/version summaries
          content:
            application/json: {schema: {$ref: '#/components/schemas/TemplatePage'}}
  /api/review/template-versions/{template_version_id}:
    get:
      tags: [Review Rules]
      operationId: getTemplateVersion
      parameters:
        - name: template_version_id
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Uuid'}
      responses:
        '200':
          description: Immutable template version
          content:
            application/json: {schema: {$ref: '#/components/schemas/TemplateVersion'}}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/review/template-versions/{template_version_id}/publish:
    post:
      tags: [Review Rules]
      operationId: publishTemplateVersion
      parameters:
        - name: template_version_id
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Uuid'}
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Published immutable template version; only published rule versions may be included
          content:
            application/json: {schema: {$ref: '#/components/schemas/TemplateVersion'}}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/review/rules:
    get:
      tags: [Review Rules]
      operationId: listRules
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: category
          in: query
          schema: {type: string}
        - name: q
          in: query
          schema: {type: string, maxLength: 200}
      responses:
        '200':
          description: Published/disabled rule versions
          content:
            application/json: {schema: {$ref: '#/components/schemas/RuleVersionPage'}}
  /api/review/rule-versions/{rule_version_id}:
    get:
      tags: [Review Rules]
      operationId: getRuleVersion
      parameters:
        - name: rule_version_id
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Uuid'}
      responses:
        '200':
          description: Immutable rule version
          content:
            application/json: {schema: {$ref: '#/components/schemas/RuleVersion'}}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/review/configurations:
    post:
      tags: [Review Rules]
      operationId: createReviewConfiguration
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/ConfigurationWrite'}}
      responses:
        '201':
          description: Configuration with pinned rule versions
          content:
            application/json: {schema: {$ref: '#/components/schemas/ReviewConfiguration'}}
        '422': {$ref: '#/components/responses/ValidationError'}
    get:
      tags: [Review Rules]
      operationId: listReviewConfigurations
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Configurations include invalid pinned rule IDs
          content:
            application/json: {schema: {$ref: '#/components/schemas/ConfigurationPage'}}
  /api/review/configurations/{configuration_id}:
    put:
      tags: [Review Rules]
      operationId: updateReviewConfiguration
      parameters:
        - name: configuration_id
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Uuid'}
        - $ref: '#/components/parameters/IfMatchRevision'
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/ConfigurationWrite'}}
      responses:
        '200':
          description: Updated configuration
          content:
            application/json: {schema: {$ref: '#/components/schemas/ReviewConfiguration'}}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/review/analysis-jobs:
    post:
      tags: [Analysis]
      operationId: startAnalysisJob
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/StartAnalysisRequest'}}
      responses:
        '202':
          description: Job accepted; same key and request replays this resource
          content:
            application/json: {schema: {$ref: '#/components/schemas/AnalysisJob'}}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
    get:
      tags: [Analysis]
      operationId: listAnalysisJobs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: batch_id
          in: query
          schema: {$ref: '#/components/schemas/Uuid'}
      responses:
        '200':
          description: Analysis job page
          content:
            application/json: {schema: {$ref: '#/components/schemas/AnalysisJobPage'}}
  /api/review/analysis-jobs/{job_id}:
    get:
      tags: [Analysis]
      operationId: getAnalysisJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Authoritative status; clients ignore responses with older revision
          content:
            application/json: {schema: {$ref: '#/components/schemas/AnalysisJob'}}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/review/analysis-jobs/{job_id}/findings:
    get:
      tags: [Analysis]
      operationId: listFindings
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: document_version_id
          in: query
          schema: {type: string}
        - name: severity
          in: query
          schema: {$ref: '#/components/schemas/Severity'}
        - name: include_suppressed
          in: query
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: Findings and current human decision overlay
          content:
            application/json: {schema: {$ref: '#/components/schemas/FindingPage'}}
  /api/review/analysis-jobs/{job_id}/retries:
    post:
      tags: [Analysis]
      operationId: retryFailedDocuments
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json: {schema: {$ref: '#/components/schemas/RetryAnalysisRequest'}}
      responses:
        '202':
          description: Child job containing failed document jobs only
          content:
            application/json: {schema: {$ref: '#/components/schemas/AnalysisJob'}}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/review/findings/{finding_id}/decision:
    put:
      tags: [Analysis]
      operationId: putFindingDecision
      parameters:
        - name: finding_id
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Uuid'}
        - $ref: '#/components/parameters/IfMatchRevision'
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/FindingDecisionWrite'}}
      responses:
        '200':
          description: Human decision; original finding remains immutable
          content:
            application/json: {schema: {$ref: '#/components/schemas/HumanDecision'}}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/review/analysis-jobs/{job_id}/decision:
    put:
      tags: [Analysis]
      operationId: putOverallDecision
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IfMatchRevision'
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/OverallDecisionWrite'}}
      responses:
        '200':
          description: Overall human decision
          content:
            application/json: {schema: {$ref: '#/components/schemas/HumanDecision'}}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/review/analysis-jobs/{job_id}/audit-events:
    get:
      tags: [Analysis]
      operationId: listAuditEvents
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Append-only event page
          content:
            application/json: {schema: {$ref: '#/components/schemas/AuditEventPage'}}
  /api/review/conversations:
    post:
      tags: [Review QA]
      operationId: createReviewConversation
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: '#/components/schemas/CreateConversationRequest'}}
      responses:
        '201':
          description: Conversation scoped to one analysis job
          content:
            application/json: {schema: {$ref: '#/components/schemas/ReviewConversation'}}
...
components:
  parameters:
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 20}
...
  responses:
...
    ValidationError:
      description: Request or business validation failed
      content:
        application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    CreateBatchRequest:
      type: object
      required: [name, declared_document_type]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
        declared_document_type: {type: string, minLength: 1, maxLength: 100}
        enhanced_ocr: {type: boolean, default: false}
...
    ReviewBatch:
      type: object
      required: [id, name, declared_document_type, enhanced_ocr, status, documents, revision, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        name: {type: string}
        declared_document_type: {type: string}
        enhanced_ocr: {type: boolean}
        status: {enum: [draft, ready, analyzing, completed, partial, failed]}
        documents: {type: array, items: {$ref: '#/components/schemas/BatchDocument'}}
        revision: {type: integer, minimum: 0}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
...
    BatchPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties: {items: {type: array, items: {$ref: '#/components/schemas/ReviewBatch'}}}
...